+++ hermes-management/src/main/java/pl/allegro/tech/hermes/management/config/kafka/KafkaConfiguration.java
@@ -34,7 +34,9 @@ import pl.allegro.tech.hermes.schema.SchemaRepository;
 import tech.allegro.schema.json2avro.converter.JsonAvroConverter;
 
 import java.time.Clock;
+import java.util.Arrays;
 import java.util.List;
+import java.util.Optional;
 import java.util.Properties;
 import java.util.stream.Collectors;
 
@@ -42,18 +44,29 @@ import static java.time.Duration.ofMillis;
 import static java.time.Duration.ofSeconds;
 import static java.util.stream.Collectors.toList;
 import static org.apache.kafka.clients.CommonClientConfigs.BOOTSTRAP_SERVERS_CONFIG;
-import static org.apache.kafka.clients.CommonClientConfigs.DEFAULT_SECURITY_PROTOCOL;
 import static org.apache.kafka.clients.CommonClientConfigs.REQUEST_TIMEOUT_MS_CONFIG;
 import static org.apache.kafka.clients.CommonClientConfigs.SECURITY_PROTOCOL_CONFIG;
 import static org.apache.kafka.common.config.SaslConfigs.SASL_JAAS_CONFIG;
 import static org.apache.kafka.common.config.SaslConfigs.SASL_MECHANISM;
-import static org.apache.kafka.common.config.SslConfigs.SSL_PROTOCOL_CONFIG;
-import static org.apache.kafka.common.config.SslConfigs.SSL_TRUSTSTORE_LOCATION_CONFIG;
-import static org.apache.kafka.common.config.SslConfigs.SSL_TRUSTSTORE_PASSWORD_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_CIPHER_SUITES_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_ENABLED_PROTOCOLS_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_ENDPOINT_IDENTIFICATION_ALGORITHM_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_ENGINE_FACTORY_CLASS_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_KEYMANAGER_ALGORITHM_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_KEYSTORE_CERTIFICATE_CHAIN_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_KEYSTORE_KEY_CONFIG;
 import static org.apache.kafka.common.config.SslConfigs.SSL_KEYSTORE_LOCATION_CONFIG;
 import static org.apache.kafka.common.config.SslConfigs.SSL_KEYSTORE_PASSWORD_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_KEYSTORE_TYPE_CONFIG;
 import static org.apache.kafka.common.config.SslConfigs.SSL_KEY_PASSWORD_CONFIG;
-import static org.apache.kafka.common.config.SslConfigs.SSL_ENDPOINT_IDENTIFICATION_ALGORITHM_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_PROTOCOL_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_PROVIDER_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_SECURE_RANDOM_IMPLEMENTATION_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_TRUSTMANAGER_ALGORITHM_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_TRUSTSTORE_CERTIFICATES_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_TRUSTSTORE_LOCATION_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_TRUSTSTORE_PASSWORD_CONFIG;
+import static org.apache.kafka.common.config.SslConfigs.SSL_TRUSTSTORE_TYPE_CONFIG;
 
 @Configuration
 @EnableConfigurationProperties(KafkaClustersProperties.class)
@@ -166,9 +179,9 @@ public class KafkaConfiguration implements MultipleDcKafkaNamesMappersFactory {
                 kafkaProperties.getKafkaConsumer().getFetchMinBytes(),
                 kafkaProperties.getKafkaConsumer().getNamePrefix(),
                 kafkaProperties.getKafkaConsumer().getConsumerGroupName(),
+                kafkaProperties.getSecurityProtocol(),
                 kafkaProperties.getSasl().isEnabled(),
                 kafkaProperties.getSasl().getMechanism(),
-                kafkaProperties.getSasl().getProtocol(),
                 kafkaProperties.getSasl().getJaasConfig(),
                 kafkaProperties.getSsl().isEnabled(),
                 kafkaProperties.getSsl().getTrustStoreLocation(),
@@ -177,9 +190,20 @@ public class KafkaConfiguration implements MultipleDcKafkaNamesMappersFactory {
                 kafkaProperties.getSsl().getKeyStorePassword(),
                 kafkaProperties.getSsl().getKeyPassword(),
                 kafkaProperties.getSsl().getProtocolVersion(),
-                kafkaProperties.getSsl().getSecurityProtocol(),
-                kafkaProperties.getSsl().getEndpointIdentificationAlgorithm()
-        );
+                kafkaProperties.getSsl().getEndpointIdentificationAlgorithm(),
+                kafkaProperties.getSsl().getKeyStoreCertificateChain(),
+                kafkaProperties.getSsl().getKeyStoreKey(),
+                kafkaProperties.getSsl().getTrustStoreCertificates(),
+                kafkaProperties.getSsl().getEnabledProtocols(),
+                kafkaProperties.getSsl().getKeyStoreType(),
+                kafkaProperties.getSsl().getProtocol(),
+                kafkaProperties.getSsl().getProvider(),
+                kafkaProperties.getSsl().getTrustStoreType(),
+                kafkaProperties.getSsl().getCipherSuites(),
+                kafkaProperties.getSsl().getEngineFactoryClass(),
+                kafkaProperties.getSsl().getKeymanagerAlgorithm(),
+                kafkaProperties.getSsl().getSecureRandomImplementation(),
+                kafkaProperties.getSsl().getTrustmanagerAlgorithm());
 
         return new KafkaConsumerPool(config, brokerStorage, configuredBootstrapServers);
     }
@@ -187,23 +211,34 @@ public class KafkaConfiguration implements MultipleDcKafkaNamesMappersFactory {
     private AdminClient brokerAdminClient(KafkaProperties kafkaProperties) {
         Properties props = new Properties();
         props.put(BOOTSTRAP_SERVERS_CONFIG, kafkaProperties.getBootstrapKafkaServer());
-        props.put(SECURITY_PROTOCOL_CONFIG, DEFAULT_SECURITY_PROTOCOL);
         props.put(REQUEST_TIMEOUT_MS_CONFIG, kafkaProperties.getKafkaServerRequestTimeoutMillis());
+        props.put(SECURITY_PROTOCOL_CONFIG, kafkaProperties.getSecurityProtocol());
+
         if (kafkaProperties.getSasl().isEnabled()) {
             props.put(SASL_MECHANISM, kafkaProperties.getSasl().getMechanism());
-            props.put(SECURITY_PROTOCOL_CONFIG, kafkaProperties.getSasl().getProtocol());
             props.put(SASL_JAAS_CONFIG, kafkaProperties.getSasl().getJaasConfig());
         }
+
         if (kafkaProperties.getSsl().isEnabled()) {
-            props.put(SASL_MECHANISM, kafkaProperties.getSsl().getMechanism());
-            props.put(SSL_TRUSTSTORE_LOCATION_CONFIG, kafkaProperties.getSsl().getTrustStoreLocation());
-            props.put(SSL_TRUSTSTORE_PASSWORD_CONFIG, kafkaProperties.getSsl().getTrustStorePassword());
+            props.put(SSL_KEY_PASSWORD_CONFIG, kafkaProperties.getSsl().getKeyPassword());
+            props.put(SSL_KEYSTORE_CERTIFICATE_CHAIN_CONFIG, kafkaProperties.getSsl().getKeyStoreCertificateChain());
+            props.put(SSL_KEYSTORE_KEY_CONFIG, kafkaProperties.getSsl().getKeyStoreKey());
             props.put(SSL_KEYSTORE_LOCATION_CONFIG, kafkaProperties.getSsl().getKeyStoreLocation());
             props.put(SSL_KEYSTORE_PASSWORD_CONFIG, kafkaProperties.getSsl().getKeyStorePassword());
-            props.put(SSL_KEY_PASSWORD_CONFIG, kafkaProperties.getSsl().getKeyPassword());
+            props.put(SSL_TRUSTSTORE_CERTIFICATES_CONFIG, kafkaProperties.getSsl().getTrustStoreCertificates());
+            props.put(SSL_TRUSTSTORE_LOCATION_CONFIG, kafkaProperties.getSsl().getTrustStoreLocation());
+            props.put(SSL_TRUSTSTORE_PASSWORD_CONFIG, kafkaProperties.getSsl().getTrustStorePassword());
+            props.put(SSL_ENABLED_PROTOCOLS_CONFIG, Optional.of(kafkaProperties.getSsl().getEnabledProtocols()).map(s -> Arrays.asList(s.split(","))).orElse(null));
+            props.put(SSL_KEYSTORE_TYPE_CONFIG, kafkaProperties.getSsl().getKeyStoreType());
             props.put(SSL_PROTOCOL_CONFIG, kafkaProperties.getSsl().getProtocolVersion());
-            props.put(SECURITY_PROTOCOL_CONFIG, kafkaProperties.getSsl().getSecurityProtocol());
+            props.put(SSL_PROVIDER_CONFIG, kafkaProperties.getSsl().getProvider());
+            props.put(SSL_TRUSTSTORE_TYPE_CONFIG, kafkaProperties.getSsl().getTrustStoreType());
+            props.put(SSL_CIPHER_SUITES_CONFIG, Optional.of(kafkaProperties.getSsl().getCipherSuites()).map(s -> Arrays.asList(s.split(","))).orElse(null));
             props.put(SSL_ENDPOINT_IDENTIFICATION_ALGORITHM_CONFIG, kafkaProperties.getSsl().getEndpointIdentificationAlgorithm());
+            props.put(SSL_ENGINE_FACTORY_CLASS_CONFIG, kafkaProperties.getSsl().getEngineFactoryClass());
+            props.put(SSL_KEYMANAGER_ALGORITHM_CONFIG, kafkaProperties.getSsl().getKeymanagerAlgorithm());
+            props.put(SSL_SECURE_RANDOM_IMPLEMENTATION_CONFIG, kafkaProperties.getSsl().getSecureRandomImplementation());
+            props.put(SSL_TRUSTMANAGER_ALGORITHM_CONFIG, kafkaProperties.getSsl().getTrustmanagerAlgorithm());
         }
 
         return AdminClient.create(props);